Getting rid of ids in game engine

+++ src/main/java/oogasalad/model/api/data/CollisionRule.java
@@ -13,6 +13,6 @@ import java.util.Map;
 @JsonPropertyOrder({"first_id", "second_id", "command"})
 public record CollisionRule(@JsonProperty("first_id") int firstId,
                             @JsonProperty("second_id") int secondId,
-                            List<Map<String, List<Double>>> command) {
+                            List<Map<String, List<Integer>>> command) {
 
 }
